notification-service/requirements.txt
boto3
python-dotenv
fastapi
uvicorn[standard]
pydantic[email]

